+++ lib/src/parse/source.dart
@@ -1,7 +1,11 @@
 /// Represents a node in Dart source code
 abstract class DartSourceNode {
+  /// The character offset into the source file
   final int offset;
+
+  /// The length of this node in characters
   final int length;
 
+  /// Create a [DartSourceNode]
   const DartSourceNode(this.offset, this.length);
 }